New Index Page + other minor changes
Revised the style and layout of the home page. Added an about section to provide some info on the purpose of the simulations. Added some new images to the main page. Added colors to the names of the wall types. Added a favicon. Changed around the wording of the instructions.