Changed File Names
There was an issue with the office speech privacy page when clicking on the Minimap. The issue only exists on the online version, not when hosted locally on my computer.

Changed the file names of he Office speech privacy files to be capital ON and OFF. The error that shows up is https://maxzking.github.io/Office%20Speech%20Privacy/Photos/Minimap/Receive1-OFF.png 404 (Not Found)